ecfr-django/pyproject.toml
[project]
name = "ecfr-django"
version = "0.1.0"
description = "eCFR Analyzer —The people voted for major reform."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-extensions>=3.2.3",
    "django-rest-framework>=0.1.0",
    "httpx>=0.28.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ipython>=8.32.0",
    "ruff>=0.9.5",
]
